Javis

Javis.draw_obj

Javis.draw_obj

Javis.draw_obj

Javis.draw_obj

Javis.float_attribute

Javis.latex

Javis.path_move

Javis.path_quadratic

Javis.pathsvg

Javis.set_attr

Javis.set_attrs

Javis.set_transform

Javis.draw_obj — Method

draw_obj(::Val{:g}, o, defs)

Draws a group by setting the attributes (like transformations) and then calls draw_obj for all child elements.

Javis.draw_obj — Method

draw_obj(::Val{:path}, o, defs)

Calls the commands specified in the path data. Currently supports only a subset of possible SVG commands.

Javis.draw_obj — Method

draw_obj(::Val{:rect}, o, defs)

Draw the rectangle defined by the object o.

Javis.draw_obj — Method

draw_obj(::Val{:use}, o, defs)

Calls the command specified in defs.

Javis.float_attribute — Method

float_attribute(o, name)

Get the attribute name of the XMLElement and parse it as a Float64

Javis.latex — Method

latex(text::LaTeXString, font_size::Real, action::Symbol)

Add the latex string text to the top left corner of the LaTeX path. Can be added to Luxor.jl graphics such as Video or Drawing.

NOTE: This only works if tex2svg is installed. It can be installed using the following command (you may have to prefix this command with sudo depending on your installation):

npm install -g mathjax-node-cli

Arguments

text::LaTeXString: a LaTeX string to render.

font_size::Real: integer font size of LaTeX string. Default 10.

action::Symbol: graphics actions defined by Luxor.jl. Default :stroke. Available actions:

:fill - See Luxor.fillpath.

:stroke - See Luxor.strokepath.

:clip - See Luxor.clip.

:fillstroke - See Luxor.fillstroke.

:fillpreserve - See Luxor.fillpreserve.

:strokepreserve - See Luxor.strokepreserve.

:none - Does nothing.

:path - See Luxor docs for polygons.md

Throws

IOError: mathjax-node-cli is not installed

Example

using Luxor
using Javis
using LaTeXStrings

my_drawing = Drawing(400, 200, "test.png")
background("white")
sethue("black")
latex(L"\sum \phi", 100)
finish()

Javis.path_move — Method

path_move(x,y)

Moving to the specified point

Javis.path_quadratic — Method

path_quadratic(c_pt::Point, x,y, xe, ye)

Drawing a quadratic bezier curve by computing a cubic one as that is supported by Luxor

Javis.pathsvg — Function

pathsvg(svg, fontsize=10)

Convert an svg to a path using Luxor. Normally called via the latex command. It handles only a subset of the full power of svg.

Javis.set_attr — Method

set_attr(::Val{:transform}, transform_str)

Call the corresponding set_transform method i.e matrix, scale and translate

Javis.set_attrs — Method

set_attrs(o)

Setting the attributes of the object o by calling set_attr methods.

Javis.set_transform — Method

set_transform(::Val{:matrix}, args...)

Multiply the new matrix with the current matrix and set it.