Javis

Javis.draw_obj

Javis.draw_obj

Javis.draw_obj

Javis.draw_obj

Javis.float_attribute

Javis.latex

Javis.path_move

Javis.path_quadratic

Javis.pathsvg

Javis.set_attr

Javis.set_attrs

Javis.set_transform

Javis.draw_obj — Method

draw_obj(::Val{:g}, o, defs)

Draws a group by setting the attributes (like transformations) and then calls draw_obj for all child elements.

Javis.draw_obj — Method

draw_obj(::Val{:path}, o, defs)

Calls the commands specified in the path data. Currently supports only a subset of possible SVG commands.

Javis.draw_obj — Method

draw_obj(::Val{:rect}, o, defs)

Draw the rectangle defined by the object o.

Javis.draw_obj — Method

draw_obj(::Val{:use}, o, defs)

Calls the command specified in defs.

Javis.float_attribute — Method

float_attribute(o, name)

Get the attribute name of the XMLElement and parse it as a Float64

Javis.latex — Method

latex(t::LaTeXString, font_size::Real, action::Symbol)

Add the latex string text at the current position. The current position is the top left corner of the latex path. This only works if tex2svg is installed. It can be installed via:

npm install -g mathjax-node-cli

The default font_size is 10 and the default action is :stroke.

Javis.path_move — Method

path_move(x,y)

Moving to the specified point

Javis.path_quadratic — Method

path_quadratic(c_pt::Point, x,y, xe, ye)

Drawing a quadratic bezier curve by computing a cubic one as that is supported by Luxor

Javis.pathsvg — Function

pathsvg(svg, fontsize=10)

Convert an svg to a path using Luxor. Normally called via the latex command. It handles only a subset of the full power of svg.

Javis.set_attr — Method

set_attr(::Val{:transform}, transform_str)

Call the corresponding set_transform method i.e matrix, scale and translate

Javis.set_attrs — Method

set_attrs(o)

Setting the attributes of the object o by calling set_attr methods.

Javis.set_transform — Method

set_transform(::Val{:matrix}, args...)

Multiply the new matrix with the current matrix and set it.