Javis

Base.:*

Javis.compute_transformation!

Javis.compute_transition!

Javis.compute_transition!

Javis.draw_obj

Javis.draw_obj

Javis.draw_obj

Javis.draw_obj

Javis.float_attribute

Javis.javis

Javis.latex

Javis.path_move

Javis.path_quadratic

Javis.pathsvg

Javis.perform_transformation

Javis.perform_transformation

Javis.perform_transformation

Javis.projection

Javis.set_attr

Javis.set_attrs

Javis.set_transform

Base.:* — Method

Base.:*(m::Array{Float64,2}, transformation::Transformation)

Convert the transformation to a matrix and multiplies m*trans_matrix. Return a new Transformation

Javis.compute_transformation! — Method

compute_transformation!(action::Action, video::Video, frame::Int)

Update action.internal_transitions for the current frame number

Javis.compute_transition! — Method

compute_transition!(internal_rotation::InternalRotation, rotation::Rotation, video, action::Action, frame)

Computes the rotation transformation for the action. If the Rotation is given directly it uses the frame number for interpolation. If rotation includes symbols the current definition of that look up is used for computation.

Javis.compute_transition! — Method

compute_transition!(internal_translation::InternalTranslation, translation::Translation, video, action::Action, frame)

Computes the translation transformation for the action. If the translation is given directly it uses the frame number for interpolation. If translation includes symbols the current definition of that look up is used for computation.

Javis.draw_obj — Method

draw_obj(::Val{:g}, o, defs)

Draws a group by setting the attributes (like transformations) and then calls draw_obj for all child elements.

Javis.draw_obj — Method

draw_obj(::Val{:path}, o, defs)

Calls the commands specified in the path data. Currently supports only a subset of possible SVG commands.

Javis.draw_obj — Method

draw_obj(::Val{:rect}, o, defs)

Draw the rectangle defined by the object o.

Javis.draw_obj — Method

draw_obj(::Val{:use}, o, defs)

Calls the command specified in defs.

Javis.float_attribute — Method

float_attribute(o, name)

Get the attribute name of the XMLElement and parse it as a Float64

Javis.javis — Method

javis(
    video::Video,
    actions::Vector{Action};
    creategif=false,
    framerate=30,
    pathname="",
    tempdirectory="",
    usenewffmpeg=true
)

Similar to animate in Luxor with a slightly different structure. Instead of using actions and a video instead of scenes in a movie.

An example:

function ground(args...) 
    background("white")
    sethue("black")
end

function circ(p=O, color="black")
    sethue(color)
    circle(p, 25, :fill)
    return Transformation(p, 0.0)
end

from = Point(-200, -200)
to = Point(-20, -130)
p1 = Point(0,-100)
p2 = Point(0,-50)
from_rot = 0.0
to_rot = 2π

demo = Video(500, 500)
javis(demo, [
    Action(1:100, ground),
    Action(1:100, :red_ball, (args...)->circ(p1, "red"), Rotation(from_rot, to_rot)),
    Action(1:100, (args...)->circ(p2, "blue"), Rotation(to_rot, from_rot, :red_ball))
], tempdirectory="images", creategif=true, pathname="rotating.gif")

This structure makes it possible to refer to positions of previous actions i.e :red_ball is an id for the position or the red ball which can be used in the Rotation of the next ball.

Javis.latex — Method

latex(text::LaTeXString, font_size::Real, action::Symbol)

Add the latex string text to the top left corner of the LaTeX path. Can be added to Luxor.jl graphics such as Video or Drawing.

NOTE: This only works if tex2svg is installed. It can be installed using the following command (you may have to prefix this command with sudo depending on your installation):

npm install -g mathjax-node-cli

Arguments

text::LaTeXString: a LaTeX string to render.

font_size::Real: integer font size of LaTeX string. Default 10.

action::Symbol: graphics actions defined by Luxor.jl. Default :stroke. Available actions:

:fill - See Luxor.fillpath.

:stroke - See Luxor.strokepath.

:clip - See Luxor.clip.

:fillstroke - See Luxor.fillstroke.

:fillpreserve - See Luxor.fillpreserve.

:strokepreserve - See Luxor.strokepreserve.

:none - Does nothing.

:path - See Luxor docs for polygons.md

Throws

IOError: mathjax-node-cli is not installed

Example

using Luxor
using Javis
using LaTeXStrings

my_drawing = Drawing(400, 200, "test.png")
background("white")
sethue("black")
latex(L"\sum \phi", 100)
finish()

Javis.path_move — Method

path_move(x,y)

Moving to the specified point

Javis.path_quadratic — Method

path_quadratic(c_pt::Point, x,y, xe, ye)

Drawing a quadratic bezier curve by computing a cubic one as that is supported by Luxor

Javis.pathsvg — Function

pathsvg(svg, fontsize=10)

Convert an svg to a path using Luxor. Normally called via the latex command. It handles only a subset of the full power of svg.

Javis.perform_transformation — Method

perform_transformation(action::Action)

Perform the transformations as described in action.internal_transitions

Javis.perform_transformation — Method

perform_transformation(trans::InternalRotation)

Translate and rotate as described in trans.

Javis.perform_transformation — Method

perform_transformation(trans::InternalTranslation)

Translate as described in trans.

Javis.projection — Method

projection(p::Point, l::Line)

Return the projection of a point to a line.

Javis.set_attr — Method

set_attr(::Val{:transform}, transform_str)

Call the corresponding set_transform method i.e matrix, scale and translate

Javis.set_attrs — Method

set_attrs(o)

Setting the attributes of the object o by calling set_attr methods.

Javis.set_transform — Method

set_transform(::Val{:matrix}, args...)

Multiply the new matrix with the current matrix and set it.