Public functions

Javis.Action — Type

Action

Defines what is drawn in a defined frame range.

Fields

frames: A range of frames for which the Action is called

id::Union{Nothing, Symbol}: An id which can be used to save the result of func

func::Function: The drawing function which draws something on the canvas. It gets called with the arguments video, action, frame

transitions::Vector{Transition} a list of transitions that can be performed before the function gets called.

internal_transitions::Vector{InternalTransition}: Similar to transitions but holds the concrete information whereas Transition can hold links to other actions which need to be computed first. See compute_transformation!

opts::Any can hold any options defined by the user

Javis.Action — Method

Action(frames, func::Function, args...)

The most simple form of an action (if there are no args/kwargs) just calls func(video, action, frame) for each of the frames it is defined for. args are defined it the next function definition and can be seen in action in this example javis

Javis.Action — Method

Action(frames, id::Union{Nothing,Symbol}, func::Function, transitions::Transition...; kwargs...)

Arguments

frames: defines for which frames this action is called

id::Symbol: Is used if the func returns something which shell be accessible by other actions later

func::Function the function that is called after the transitions are performed

transitions::Transition... a list of transitions that are performed before the function func itself is called

The keywords arguments will be saved inside .opts as a Dict{Symbol, Any}

Javis.Line — Type

Line

A type to define a line by two points. Can be used i.e. in projection We mean the mathematic definition of a continuous line and not a segment of a line.

Fields

p1::Point: start point

p2::Point: second point to define the line)

Javis.Rotation — Type

Rotation <: Transition

Stores the rotation similar to Translation with from and to but also the rotation point.

Fields

from::Union{Float64, Symbol}: The start rotation or a link to it

to::Union{Float64, Symbol}: The end rotation or a link to it

center::Union{Point, Symbol}: The center of the rotation or a link to it.

Javis.Rotation — Method

Rotation(from, to)

Rotation as a transition from from to to (in radians) around the origin.

Javis.Transformation — Type

Transformation

Defines a transformation which can be returned by an action to be accessible later. See the circ function inside the javis as an example. It can be accessed by another [Action])(@ref) using the symbol notation like :red_ball in the example.

Fields

p::Point: the translation part of the transformation

angle::Float64: the angle component of the transformation (in radians)

Javis.Translation — Type

Translation <: Transition

Stores the Point or a link for the start and end position of the translation

Fields

from::Union{Point, Symbol}: The start position or a link to the start position. See :red_ball in javis to::Union{Point, Symbol}: The end position or a link to the end position

Javis.Video — Type

Video

Defines the video canvas for an animation.

Fields

width::Int the width in pixel

height::Int the height in pixel

defs::Dict{Symbol, Any} Some definitions which should be accessible throughout the video.

Javis.Video — Method

Video(width, height)

Create a video with a certain width and height in pixel.

Javis.BackgroundAction — Method

BackgroundAction(frames, func::Function, args...; kwargs...)

Create an Action where in_global_layer is set to true such that i.e the specified color in the background is applied globally (basically a new default)

Javis.javis — Method

javis(
    video::Video,
    actions::Vector{ActionType};
    creategif=false,
    framerate=30,
    pathname="",
    tempdirectory="",
    usenewffmpeg=true
)

Similar to animate in Luxor with a slightly different structure. Instead of using actions and a video instead of scenes in a movie.

Arguments

video::Video: The video which defines the dimensions of the output

actions::Vector{Action}: All actions that are performed

Keywords

creategif::Bool: defines whether the images should be rendered to a gif

framerate::Int: The frame rate of the video

pathname::String: The path for the gif if creategif = true

tempdirectory::String: The folder where each frame is stored

deletetemp::Bool: If true and creategif is true => tempdirectory is emptied after the gif is created

Example

function ground(args...) 
    background("white")
    sethue("black")
end

function circ(p=O, color="black")
    sethue(color)
    circle(p, 25, :fill)
    return Transformation(p, 0.0)
end

from = Point(-200, -200)
to = Point(-20, -130)
p1 = Point(0,-100)
p2 = Point(0,-50)
from_rot = 0.0
to_rot = 2π

demo = Video(500, 500)
javis(demo, [
    Action(1:100, ground),
    Action(1:100, :red_ball, (args...)->circ(p1, "red"), Rotation(from_rot, to_rot)),
    Action(1:100, (args...)->circ(p2, "blue"), Rotation(to_rot, from_rot, :red_ball))
], tempdirectory="images", creategif=true, pathname="rotating.gif")

This structure makes it possible to refer to positions of previous actions i.e :red_ball is an id for the position or the red ball which can be used in the Rotation of the next ball.

Javis.latex — Method

latex(text::LaTeXString, font_size::Real, action::Symbol)

Add the latex string text to the top left corner of the LaTeX path. Can be added to Luxor.jl graphics such as Video or Drawing.

NOTE: This only works if tex2svg is installed. It can be installed using the following command (you may have to prefix this command with sudo depending on your installation):

npm install -g mathjax-node-cli

Arguments

text::LaTeXString: a LaTeX string to render.

font_size::Real: integer font size of LaTeX string. Default 10.

action::Symbol: graphics actions defined by Luxor.jl. Default :stroke. Available actions:

:fill - See Luxor.fillpath.

:stroke - See Luxor.strokepath.

:clip - See Luxor.clip.

:fillstroke - See Luxor.fillstroke.

:fillpreserve - See Luxor.fillpreserve.

:strokepreserve - See Luxor.strokepreserve.

:none - Does nothing.

:path - See Luxor docs for polygons.md

Throws

IOError: mathjax-node-cli is not installed

Example

using Luxor
using Javis
using LaTeXStrings

my_drawing = Drawing(400, 200, "test.png")
background("white")
sethue("black")
latex(L"\sum \phi", 100)
finish()

Javis.morph — Method

morph(from_func::Function, to_func::Function)

A closure for the morph_internal function. This makes it easier to write the function inside an Action.

Currently morphing is quite simple and only works for basic shapes. It especially does not work with functions which produce more than one polygon or which produce filled polygons. Blending between fills of polygons is definitely coming at a later stage.

Important: The functions itself should not draw the polygon i.e. use circle(Point(100,100), 50) instead of circle(Point(100,100), 50, :stroke)

Arguments

from_func::Function: The function that creates the path for the first polygon.

to_func::Function: Same as from_func but it defines the "result" polygon, which will be displayed at the end of the Action

Example

This creates a star that morphs into a circle and back.

using Luxor
using Javis

astar(args...) = star(O, 50) 
acirc(args...) = circle(Point(100,100), 50) 

video = Video(500, 500)
javis(video, [
    Action(1:100, ground),
    Action(1:50, morph(astar, acirc)),
    Action(51:100, morph(acirc, astar))
], creategif=true, tempdirectory="images", 
    pathname="star2circle.gif", deletetemp=true)

Javis.projection — Method

projection(p::Point, l::Line)

Return the projection of a point to a line.

Private functions

Base.:* — Method

Base.:*(m::Array{Float64,2}, transformation::Transformation)

Convert the transformation to a matrix and multiplies m*trans_matrix. Return a new Transformation

Javis.compute_transformation! — Method

compute_transformation!(action::ActionType, video::Video, frame::Int)

Update action.internal_transitions for the current frame number

Javis.compute_transition! — Method

compute_transition!(internal_rotation::InternalRotation, rotation::Rotation, video, action::ActionType, frame)

Computes the rotation transformation for the action. If the Rotation is given directly it uses the frame number for interpolation. If rotation includes symbols the current definition of that look up is used for computation.

Javis.compute_transition! — Method

compute_transition!(internal_translation::InternalTranslation, translation::Translation, video, action::ActionType, frame)

Computes the translation transformation for the action. If the translation is given directly it uses the frame number for interpolation. If translation includes symbols the current definition of that look up is used for computation.

Javis.draw_obj — Method

draw_obj(::Val{:g}, o, defs)

Draws a group by setting the attributes (like transformations) and then calls draw_obj for all child elements.

Javis.draw_obj — Method

draw_obj(::Val{:path}, o, defs)

Calls the commands specified in the path data. Currently supports only a subset of possible SVG commands.

Javis.draw_obj — Method

draw_obj(::Val{:rect}, o, defs)

Draw the rectangle defined by the object o.

Javis.draw_obj — Method

draw_obj(::Val{:use}, o, defs)

Calls the command specified in defs.

Javis.float_attribute — Method

float_attribute(o, name)

Get the attribute name of the XMLElement and parse it as a Float64

Javis.match_num_point! — Method

match_num_point!(poly_1::Vector{Point}, poly_2::Vector{Point})

This is a helper function for morph. Given two polygons poly_1 and poly_2 points are added to the polygon with less points until both polygons have the same number of points. The polygon with less points gets mutated during this process.

Arguments

poly_1::Vector{Point}: The points which define the first polygon

poly_2::Vector{Point}: The points which define the second polygon

Javis.morph_internal — Method

morph_internal(video::Video, action::Action, frame, from_func::Function, to_func::Function)

Internal version of morph but described there.

Javis.path_move — Method

path_move(x,y)

Moving to the specified point

Javis.path_quadratic — Method

path_quadratic(c_pt::Point, x,y, xe, ye)

Drawing a quadratic bezier curve by computing a cubic one as that is supported by Luxor

Javis.pathsvg — Function

pathsvg(svg, fontsize=10)

Convert an svg to a path using Luxor. Normally called via the latex command. It handles only a subset of the full power of svg.

Javis.perform_action — Method

perform_action(action, video, frame, origin_matrix)

Is called inside the javis and does everything handled for an ActionType. It is a 4-step process:

compute the transformation for this action (translation, rotation, scale)

do the transformation

call the action function

save the result of the action if wanted inside video.defs

Javis.perform_transformation — Method

perform_transformation(action::ActionType)

Perform the transformations as described in action.internal_transitions

Javis.perform_transformation — Method

perform_transformation(trans::InternalRotation)

Translate and rotate as described in trans.

Javis.perform_transformation — Method

perform_transformation(trans::InternalTranslation)

Translate as described in trans.

Javis.save_morph_polygons! — Method

save_morph_polygons!(action::Action, from_func::Function, to_func::Function)

Converts the paths created by the functions to polygons and calls match_num_point! such that both polygons have the same number of points. This is done once inside morph_internal. Saves the two polygons inside action.opts[:from_poly] and action.opts[:to_poly].

Assumption: Both functions create only a single polygon each.

Javis.set_attr — Method

set_attr(::Val{:transform}, transform_str)

Call the corresponding set_transform method i.e matrix, scale and translate

Javis.set_attrs — Method

set_attrs(o)

Setting the attributes of the object o by calling set_attr methods.

Javis.set_transform — Method

set_transform(::Val{:matrix}, args...)

Multiply the new matrix with the current matrix and set it.