Public functions

Javis.Action — Type

Action <: AbstractAction

A Action can be used in the keyword arguments of an Object to define small sub objects on the object function, such as appear.

A Action should not be created by hand but instead by using one of the constructors.

Fields

frames::Frames: the frames relative to the parent Object

anim::Animation: defines the interpolation function for the transition

func::Function: the function that gets called in each of those frames. Takes the following arguments: video, object, action, rel_frame

transition::Union{Nothing, AbstractTransition}

keep::Bool defines whether this Action is called even after the last frame it was defined on

defs::Dict{Symbol, Any} any kind of definitions that are relevant for the action.

Javis.Action — Method

Action([frames], [Animation], func::Function; keep=true)

An Action gives an Object the opportunity to move, change color or much more. It can be defined in many different ways.

Arguments

frames can be a Symbol, a UnitRange or a Glob to define them in a global way.

Default: If not defined it will be the same as the previous Action or if it's the first action then it will be applied for the whole length of the object.

It defines for which frames the action acts on the object.

These are defined in a relative fashion so 1:10 means the first ten frames of the object and not the first ten frames of the Video

animation can be an easing function or animation which can be defined by Animations.jl

Default: The default is linear()

Possible simple easing functions is sineio() for more check Animations.jl

func is the function that describes the actual action

It can be a general function which takes in the following four arguments

video, object, action, rel_frame

If you don't need them you can write (args...)->your_function(arg1, arg2)

You often don't need an own function and instead can use predefined functions like

appear, disappear, follow_path

Keywords

keep::Bool defaults to true defines whether the Action is called even for frames after it's last defined. In more simple terms: If one has Action(1:10, anim, translate()) It will get translated to the last position on frame 11:END_OF_OBJECT. One can set ; keep = false to turn off this behavior.

Example

function ground(args...)
    background("black")
    sethue("white")
end

video = Video(500, 500)
Background(1:100, ground)
obj = Object((args...)->circle(O, 50, :fill))
act!(obj, Action(1:20, appear(:fade)))
act!(obj, Action(21:50, Translation(50, 50)))
act!(obj, Action(51:80, Translation(-50, -50)))
act!(obj, Action(81:100, disappear(:fade)))
render(video; pathname="test.gif")

Javis.Glob — Type

Glob

Ability to define frames in a gloabl fashion inside Action.

Example

red_circ = Object(1:90, (args...)->circ("red"))
blue_circ = Object(21:90, (args...)->circ("blue"))
act!([red_circ, blue_circ], Action(Glob(85:90), disappear(:fade)))

is the same as

red_circ = Object(1:90, (args...)->circ("red"))
blue_circ = Object(20:90, (args...)->circ("blue"))
act!(red_circ, Action(85:90, disappear(:fade)))
act!(blue_circ, Action(65:70), disappear(:fade))

Fields

frames::UnitRange defines the frames in a global fashion.

Javis.Line — Type

Line

A type to define a line by two points. Can be used i.e. in projection We mean the mathematic definition of a continuous line and not a segment of a line.

Fields

p1::Point: start point

p2::Point: second point to define the line

Javis.Object — Type

Object

Defines what is drawn in a defined frame range.

Fields

frames::Frames: A range of frames for which the Object is called

func::Function: The drawing function which draws something on the canvas. It gets called with the arguments video, object, frame

start_pos::Union{Object, Point} defines the origin of the object. It gets translated to this point

actions::Vector{AbstractAction} a list of actions applied to this object

current_setting:: The current state of the object see ObjectSetting

opts::Any can hold any options defined by the user

change_keywords::Dict{Symbol,Any} the modified keywords changed by change

result::Vector the result of the object (if something gets returned)

Javis.Object — Method

Object([frames], func::Function, [start_pos]; kwargs...)

Arguments

frames can be a Symbol, a UnitRange or a relative way to define frames see Rel

Default: If not defined it will be the same as the previous Object.

Important: The first Object needs the frames specified as a UnitRange.

It defines for which frames the object is active

func is a Function and the only required argument

This defines the actual object that gets drawn.

The function takes the following three arguments:

video, object, frame

If you don't need them you can write (args...)->your_function(arg1, arg2)

Example

function ground(args...)
    background("black")
    sethue("white")
end

video = Video(500, 500)
Background(1:100, ground)
Object((args...)->circle(O, 50, :fill))
render(video; pathname="test.gif")

Here the Background uses the named way of defining the function whereas the circle object is defined in the anonymous function (args...)->circle(O, 50, :fill). It basically depends whether you want to have a simple Luxor object or something more complex.

Javis.Rel — Type

Rel

Ability to define frames in a relative fashion.

Example

Background(1:100, ground)
Object(1:90, (args...)->circ("red"))
Object(Rel(10), (args...)->circ("blue"))
Object((args...)->circ("red"))

is the same as

Background(1:100, ground)
Object(1:90, (args...)->circ("red"))
Object(91:100, (args...)->circ("blue"))
Object(91:100, (args...)->circ("red"))

Fields

frames::UnitRange defines the frames in a relative fashion.

Javis.Rel — Method

Rel(i::Int)

Shorthand for Rel(1:i)

Javis.Transformation — Type

Transformation

Defines a transformation which can be returned by an object to be accessible later. See the circ function inside the javis as an example.

It can be accessed by another [Object])(@ref) using the symbol notation like :red_ball in the example.

Fields

point::Point: the translation part of the transformation

angle::Float64: the angle component of the transformation (in radians)

scale::Tuple{Float64, Float64}: the scaling component of the transformation

Javis.Video — Type

Video

Defines the video canvas for an animation.

Fields

width::Int the width in pixel

height::Int the height in pixel

objects::Vector{AbstractObject} the objects defined in this video

defs::Dict{Symbol, Any} Some definitions which should be accessible throughout the video.

Javis.Video — Method

Video(width, height)

Create a video with a certain width and height in pixel. This also sets CURRENT_VIDEO.

Base.:* — Method

Base.:*(m::Array{Float64,2}, transformation::Transformation)

Convert the transformation to a matrix and multiplies m*trans_matrix. Return a new Transformation

Javis.Background — Method

Background(frames, func)

The Background is internally just an Object and can be defined the same way. In contrast to an object this a Background will change the global canvas and not just a layer. Normally it's used to define defaults and the background color. See Luxor.background

Example

function ground(args...)
    background("black")
    sethue("white")
end

video = Video(500, 500)
Background(1:100, ground)
Object((args...)->circle(O, 50, :fill))
render(video; pathname="test.gif")

This draws a white circle on a black background as sethue is defined for the global frame.

Javis.ang — Method

ang(x)

ang is just a short-hand for get_angle

Javis.appear — Method

appear(s::Symbol)

Appear can be used inside an Action to make an Object appear.

Example

house = Object(101:200, (args...)->house_of_nicholas())
act!(house, Action(1:20, appear(:fade)))
act!(house, Action(81:100, disappear(:fade)))

In this case the house_of_nicholas will fade in during the first 20 frames of the Object so 101-120.

Arguments

s::Symbol: the symbol defines the animation of appearance The only symbols that are currently supported are:

:fade_line_width which increases the line width up to the default value or the value specified by setline

:fade which increases the opcacity up to the default value or the value specified by setopacity

:scale which increases the scale up to the default value 1 or the value specified by scale

:draw_text which only works for text and lets it appear from left to right.

Javis.change — Method

change(s::Symbol, [vals::Pair])

Changes the keyword s of the parent Object from vals[1] to vals[2] in an animated way.

Arguments

s::Symbol Change the keyword with the name s

vals::Pair If vals is given i.e 0 => 25 it will be animated from 0 to 25.

The default is to use 0 => 1 or use the value given by the animation

defined in the SubAction

Example

Background(1:100, ground)
obj = Object((args...; radius = 25) -> object(O, radius, "red"), Point(100, 0))
act!(obj, Action(1:50, change(:radius, 25 => 0)))
act!(Action(51:100, change(:radius, 0 => 25)))

Javis.disappear — Method

disappear(s::Symbol)

Disappear can be used inside an Action to make an Object disappear.

Example

house = Object(101:200, (args...)->house_of_nicholas())
act!(house, Action(1:20, appear(:fade)))
act!(house, Action(81:100, disappear(:fade)))

In this case the house_of_nicholas will fade out during the last 20 frames of the Object so 181-200.

Arguments

s::Symbol: the symbol defines the animation of disappearance The only symbols that are currently supported are:

:fade_line_width which decreases the line width down to 0

:fade which decreases the opacity down to 0

:scale which decreases the scale down to 0

:draw_text which only works for text and let the text disappear from right to left.

Javis.draw_grid — Method

draw_grid(video::Video, object::AbstractObject, frame::Int; direction::AbstractString = "TR", line_gap = 25)

Draws an oriented grid on the given frame of a Video.

Arguments

direction::AbstractString: Where grid animation finishes. Default: "TR" Available Orientations:

"TR" - Animation finishes in the Top Right corner of the frame.

"TL" - Animation finishes in the Top Left corner of the frame.

"BR" - Animation finishes in the Bottom Right corner of the frame.

"BL" - Animation finishes in the Bottom Left corner of the frame.

line_gap: How many pixels between each line. Default: 25

Example

Example call of this function within an Object.

...
 Object(1:100, :line, draw_grid(direction = "TL", line_gap = 25))
...

Javis.follow_path — Method

follow_path(points::Vector{Point}; closed=true)

Can be applied inside an action such that the parent object follows a path. It takes a vector of points which can be created as an example by calling circle(O, 50). Notice that the object is set to :none, the default.

Example

Action(1:150, follow_path(star(O, 300)))

Arguments

points::Vector{Point} - the vector of points the object should follow

Keywords

closed::Bool default: true, sets whether the path is a closed path as for example when using a circle, ellipse or any polygon. For a bezier path it should be set to false.

Javis.fontsize — Method

fontsize(fsize)

Same as Luxor.fontsize: Sets the current font size.

Example

fontsize(12)
text("Hello World!")

Arguments:

fsize: the new font size

Javis.get_angle — Method

get_angle(obj::AbstractObject)

Get access to the angle that got saved in a previous object.

Returns

Float64: the angle stored by a previous object i.e via return Transformation(p, angle)

Javis.get_fontsize — Method

get_fontsize(fsize)

Same as Luxor.get_fontsize but works with every version of Luxor that is supported by Javis.

Example

fontsize(12)
fsize = get_fontsize()
text("Hello World! $fsize")

Returns

Float64: the current font size

Javis.get_position — Method

get_position(obj::Object)

Get access to the position that got saved in a previous object.

Returns

Point: the point stored by a previous object.

Throws

If the function of Object didn't return a Point or Transformation

Javis.get_scale — Method

get_scale(obj::AbstractObject)

Get access to the scaling that got saved in a previous object.

Returns

Scaling: the scale stored by a previous object.

Javis.latex — Method

latex(text::LaTeXString, pos::Point, object::Symbol)

Add the latex string text to the top left corner of the LaTeX path. Can be added to Luxor.jl graphics via Video.

NOTES:

This only works if tex2svg is installed. It can be installed using the following command (you may have to prefix this command with sudo depending on your installation):

  npm install -g mathjax-node-cli

The latex method must be called from within an Object.

Arguments

text::LaTeXString: a LaTeX string to render.

pos::Point: position of the upper left corner of the latex text. Default: O

can be written as x, y instead of Point(x, y)

object::Symbol: graphics objects defined by Luxor.jl. Default :stroke.

Available objects:

:stroke - Draws the latex string on the canvas. For more info check Luxor.strokepath

:path - Creates the path of the latex string but does not render it to the canvas.

Throws

IOError: mathjax-node-cli is not installed

Example

using Javis
using LaTeXStrings

function ground(args...)
    background("white")
    sethue("black")
end

function draw_latex(video, object, frame)
    fontsize(50)
    x = 100
    y = 120
    latex(L"\sqrt{5}", x, y)
end

demo = Video(500, 500)
Background(1:2, ground)
Object(draw_latex)
render(demo; pathname = "latex.gif")

Javis.morph_to — Method

morph_to(to_func::Function; object=:stroke)

A closure for the _morph function. This makes it easier to write the function inside an Object.

Currently morphing is quite simple and only works for basic shapes. It especially does not work with functions which produce more than one polygon or which produce filled polygons. Blending between fills of polygons is definitely coming at a later stage.

Important: The functions itself should not draw the polygon i.e. use circle(Point(100,100), 50) instead of circle(Point(100,100), 50, :stroke)

Arguments

to_func::Function: Same as from_func but it defines the "result" polygon, which will be displayed at the end of the Object

Keywords

object::Symbol defines whether the object has a fill or just a stroke. Defaults to stroke.

Example

This creates a star that morphs into a circle and back.

astar(args...) = star(O, 50)
acirc(args...) = circle(Point(100,100), 50)

video = Video(500, 500)
back = Background(1:20, ground)
star_obj = Object(1:10, astar)
act!(star_obj, Action(linear(), morph_to(acirc)))
circle_obj = Object(11:20, acirc)
act!(circle_obj, Action(:same, morph_to(astar)))

Javis.pos — Method

pos(x)

pos is just a short-hand for get_position

Javis.projection — Method

projection(p::Point, l::Line)

Return the projection of a point to a line.

Javis.render — Method

render(
    video::Video;
    framerate=30,
    pathname="javis_GIBBERISH.gif",
    tempdirectory="",
    liveview=false
)

Renders all previously defined Object drawings to the user-defined Video as a gif or mp4.

Arguments

video::Video: The video which defines the dimensions of the output

Keywords

framerate::Int: The frame rate of the video

pathname::String: The path for the rendered gif or mp4 (i.e output.gif or output.mp4)

Default: The animation is rendered as a gif with the javis_ prefix and some gibberish afterwards

tempdirectory::String: The folder where each frame is stored Defaults to a temporary directory when not set

liveview::Bool: Causes a live image viewer to appear to assist with animation development

Javis.rev — Method

rev(e::Easing)

Reverse an easing function such that easing_to_animation maps it to [1.0, 0.0] instead of [0.0, 1.0]. An example can be seen in rotate

Javis.rotate_around — Method

rotate_around(p)

Rotate an Object using an Action and an Animation defined with Animations.jl around a point p. For rotate it rotates around the current origin.

An example can be seen in rotate.

Arguments

p: the point to rotate around

Javis.scale — Method

scale(scl_x, scl_y)

Same as scale but the x scale and y scale can be changed independently.

Arguments:

scl_x: scale in x direction

scl_y: scale in y direction

Javis.scale — Method

scale(scl)

Set the scale and multiply it with the current multiplier which is i.e. set by appear and disappear.

Normal behavior without any animation is the same as Luxor.scale.

Example

scale(0.5)
circle(O, 20, :fill) # the radius would be 10 because of the scaling

Arguments:

scl: the new default scale

Javis.scale — Method

scale()

Scale a function defined inside a Action using an Animation defined with Animations.jl.

An example can be seen in rotate.

Javis.scaleto — Method

scaleto(x, y)

Scale to a specific scaling instead of multiplying it with the current scale. For scaling on top of the current scale have a look at scale.

Javis.scl — Method

scl(x)

scl is just a short-hand for get_scale

Javis.setline — Method

setline(linewidth)

Set the line width and multiply it with the current multiplier which is i.e. set by appear and disappear.

Normal behavior without any animation is the same as Luxor.setline.

Example

setline(10)
line(O, Point(10, 10))

Arguments:

linewidth: the line width in pixel

Javis.setopacity — Method

setopacity(opacity)

Set the opacity and multiply it with the current multiplier which is i.e. set by appear and disappear.

Normal behavior without any animation is the same as Luxor.setopacity.

Example

setopacity(0.5)
circle(O, 20, :fill)

Arguments:

opacity: the opacity between 0.0 and 1.0

Javis.setopacity — Method

setopacity()

Set the color of an Object using an Action and an Animation defined with Animations.jl.

Example

A possible animation would look like this:

opacity_anim = Animation(
    [0, 0.5, 1], # must go from 0 to 1
    [
        0.0,
        0.3,
        0.7,
    ],
    [sineio(), sineio()],
)

An example on how to integrate this into an Action can be seen in rotate. Where this would be a valid Action: Action(1:150, opacity_anim, setopacity()).

Javis.text — Function

text(str, pos = O; valign = :baseline, halign = :left, angle = 0.0)

Has bacially the same functionality as Luxor.text but overrides that method to allow to animate text with appear.

Example

text_obj = Object(1:100, (args...) -> text("Hello Stream!"; halign = :center))
act!(text_obj, Action(1:15, sineio(), appear(:draw_text)))
act!(text_obj, Action(76:100, sineio(), disappear(:draw_text)))

draws the text from left to right in the first 15 frames and in the last 15 frames it disappears.

Arguments

str::AbstractString the string that should be shown

pos::Point defaults to the origin and can be written as x,y as well as Point(x,y).

Keywords

valign::Symbol defaults to :baseline and takes (:top, :middle, :bottom, :baseline)

halign::Symbol defaults to :left and takes (:left, :center, :centre, :right)

angle::Float64 defaults to 0.0 and specifies the angle of the text

Javis.val — Method

val(x)

val is just a short-hand for get_value

Javis.zero_lines — Method

zero_lines(video::Video, object::AbstractObject, frame::Int; direction::AbstractString = "TR", line_thickness = 10)

Draws zero lines on the given frame of a Video.

Arguments

direction::AbstractString: Direction for how vertical and horizontal axes are drawn.

Default: "TR" Available Orientations:

"TR" - Vertical axis drawn towards the Top and horizontal axis drawn to the Right of the frame.

"TL" - Vertical axis drawn towards the Top and horizontal axis drawn to the Left of the frame.

"BR" - Vertical axis drawn towards the Bottom and horizontal axis drawn to the Right of the frame.

"BL" - Vertical axis drawn towards the Bottom and horizontal axis drawn to the Left of the frame.

line_thickness: Defines the thickness of the zero lines. Default: 10

Example

This example will produce an animation with the vertical axis being drawn towards the top and the horizontal axis being drawn towards the left. One will need to define their own path for tempdirectory and pathname.

...
 Object(1:100, :line, zero_lines(direction = "TL", line_thickness = 10)),
...

Luxor.rotate — Method

rotate()

Rotate an Object using an Action and an Animation defined with Animations.jl.

If you're used to working with Animations.jl this should feel quite natural. Instead of defining each movement in its own action it's possible to define it in one by using an Animation.

Example

using Javis, Animations

# define ground function here

video = Video(500, 500)
translate_anim = Animation(
    [0, 1], # must go from 0 to 1
    [O, Point(150, 0)],
    [sineio()],
)

translate_back_anim = Animation(
    [0, 1], # must go from 0 to 1
    [O, Point(-150, 0)],
    [sineio()],
)

rotate_anim = Animation(
    [0, 1], # must go from 0 to 1
    [0, 2π],
    [linear()],
)

Background(1:150, ground)
ball = Object((args...) -> circle(O, 25, :fill))
act!(ball, Action(1:10, sineio(), scale()))
act!(ball, Action(11:50, translate_anim, translate()))
act!(ball, Action(51:100, rotate_anim, rotate_around(Point(-150, 0))))
act!(ball, Action(101:140, translate_back_anim, translate()))
act!(ball, Action(141:150, rev(sineio()), scale()))

render(video)

which uses the Action syntax five times with both easing functions directly and animation objects. The rev(sineio()) creates an Animation which goes from 1.0 to 0.0.

Luxor.sethue — Method

sethue()

Set the color of an Object using an Action and an Animation defined with Animations.jl.

Example

A possible animation would look like this:

color_anim = Animation(
    [0, 0.5, 1], # must go from 0 to 1
    [
        Lab(colorant"red"),
        Lab(colorant"cyan"),
        Lab(colorant"black"),
    ],
    [sineio(), sineio()],
)

An example on how to integrate this into an Action can be seen in rotate. Where this would be a valid Action: Action(1:150, color_anim, sethue()).

Luxor.translate — Method

translate()

Translate an Object using an Action and an Animation defined with Animations.jl.

If you're used to working with Animations.jl this should feel quite natural. Instead of defining each movement in its own action it's possible to define it in one by using an Animation.

Example

using Javis, Animations

function ground(args...)
    background("black")
    sethue("white")
end

video = Video(500, 500)
circle_anim = Animation(
    [0.0, 0.3, 0.6, 1.0], # must go from 0 to 1
    # the circle will move from the origin to `Point(150, 0)` then `Point(150, 150)`
    # and back to the origin `O`.
    [O, Point(150, 0), Point(150, 150), O],
    [sineio(), polyin(5), expin(8)],
)

Background(1:150, ground)
obj = Object((args...)->circle(O, 25, :fill))
act!(obj, Action(1:150, circle_anim, translate()))

render(video)

This notation uses the Animations.jl library very explicitly. It's also possible to do the same with:

obj = Object((args...)->circle(O, 25, :fill))
act!(obj, Action(1:50, sineio(), Translation(150, 0)))
act!(obj, Action(51:100, polyin(2), Translation(0, 150)))
act!(obj, Action(101:150, expin(8), Translation(-150, -150)))

which uses the Action syntax three times and only uses easing functions instead of specifying the Animation directly.

Here circle_anim defines the movement of the circle. The most important part is that the time in animations has to be from 0.0 to 1.0.

Private functions

Javis.CURRENT_OBJECT — Constant

CURRENT_OBJECT

holds the current object in an array to be declared as a constant The current object can be accessed using CURRENT_OBJECT[1]

Javis.CURRENT_VIDEO — Constant

CURRENT_VIDEO

holds the current video in an array to be declared as a constant The current video can be accessed using CURRENT_VIDEO[1]

Javis.Frames — Type

Frames

Stores the actual computed frames and the user input which can be :same or Rel(10). The frames are computed in render.

Javis.ObjectSetting — Type

ObjectSetting

The current settings of an Object which are saved in object.current_setting.

Fields

line_width::Float64: the current line width

mul_line_width::Float64: the current multiplier for line width. The actual line width is then: mul_line_width * line_width

opacity::Float64: the current opacity

mul_opacity::Float64: the current multiplier for opacity. The actual opacity is then: mul_opacity * opacity

fontsize::Float64 the current font size

show_object::Bool is set to false if scale would be 0.0 which is forbidden by Cairo

current_scale::Tuple{Float64, Float64}: the current scale

desired_scale::Tuple{Float64, Float64}: the new desired scale

mul_scale::Float64: the multiplier for the new desired scale. The actual new scale is then: mul_scale * desired_scale

Javis.ReversedEasing — Type

ReversedEasing

Will be used to reverse an easing inside easing_to_animation. Can be constructed from an easing function using rev.

Javis._decrement — Method

_decrement(video::Video, widgets::Vector, objects::Vector, dims::Vector,
    canvas::Gtk.Canvas, frames::Int)

Decrements a given value and returns the associated frame.

Javis._draw_image — Method

_draw_image(video::Video, objects::Vector, frame::Int, canvas::Gtk.Canvas,
img_dims::Vector)

Internal function to create an image that is drawn on a Gtk Canvas.

Javis._increment — Method

_increment(video::Video, widgets::Vector, objects::Vector, dims::Vector,
    canvas::Gtk.Canvas, frames::Int)

Increments a given value and returns the associated frame.

Javis._javis_viewer — Function

 _javis_viewer(video::Video, frames::Int, object_list::Vector, show::Bool)

Internal Javis Viewer built on Gtk that is called for live previewing.

Javis._morph_to — Method

_morph_to(video::Video, object::AbstractObject, frame, from_func::Function, to_func::Function; draw_object=:stroke)

Internal version of morph but described there.

Javis.animate_text — Method

animate_text(
    str,
    pos::Point,
    valign::Symbol,
    halign::Symbol,
    angle::Float64,
    t::Float64,
)

This function is used as a subfunction of text and animates the str by clipping the textoutlines and creating a growing circle in the lower left corner to display the text from left to right in an animated fashion.

Javis.compute_frames! — Method

compute_frames!(elements::Vector{UA}; parent=nothing)
    where UA<:Union{AbstractObject,AbstractAction}

Set elem.frames.frames to the computed frames for each elem in elements.

Javis.draw_obj — Method

draw_obj(::Val{:g}, o, defs)

Draws a group by setting the attributes (like transformations) and then calls draw_obj for all child elements.

Javis.draw_obj — Method

draw_obj(::Val{:path}, o, defs)

Calls the commands specified in the path data. Currently supports only a subset of possible SVG commands.

Javis.draw_obj — Method

draw_obj(::Val{:rect}, o, defs)

Draw the rectangle defined by the object o.

Javis.draw_obj — Method

draw_obj(::Val{:use}, o, defs)

Calls the command specified in defs.

Javis.draw_object — Method

draw_object(object, video, frame, origin_matrix)

Is called inside the javis and does everything handled for an AbstractObject. It is a 4-step process:

translate to the start position

call the relevant actions

call the object function

save the result of the object if wanted inside video.defs

Javis.easing_to_animation — Method

easing_to_animation(easing)

Converts an easing to an Animation with time goes from 0.0 to 1.0 and value from 0 to 1.

Javis.easing_to_animation — Method

easing_to_animation(rev_easing::ReversedEasing)

Converts an easing to an Animation with time goes from 0.0 to 1.0 and value from 1 to 0.

Javis.float_attribute — Method

float_attribute(o, name)

Get the attribute name of the XMLElement and parse it as a Float64

Javis.get_current_setting — Method

get_current_setting()

Return the current setting of the current object

Javis.get_frames — Method

get_frames(parent, elem, glob::Glob, last_frames::UnitRange)

Return the frames based on a global frames Glob object and the last_frames. If is_action is false this is the same as defining the frames as just a unit range. Inside an action it's now defined globally though.

Javis.get_frames — Method

get_frames(parent, elem, relative::Rel, last_frames::UnitRang; is_first=falsee)

Return the frames based on a relative frames Rel object and the last_frames.

Javis.get_frames — Method

get_frames(parent, elem, frames::Symbol, last_frames::UnitRange; is_first=false)

Get the frames based on a symbol (currently only same) and the last_frames. Throw ArgumentError if symbol is unknown

Javis.get_frames — Method

get_frames(elem)

Return elem.frames.frames which holds the computed frames for the AbstractObject or AbstractAction a.

Javis.get_interpolation — Method

get_interpolation(action::AbstractAction, frame)

Return the value of the action.anim Animation based on the relative frame given by get_interpolation(get_frames(action), frame)

Javis.get_interpolation — Method

get_interpolation(frames::UnitRange, frame)

Return a value between 0 and 1 which represents the relative frame inside frames.

Javis.get_javis_frame — Method

get_javis_frame(video, objects, frame)

Get a frame from an animation given a video object, its objects, and frame.

Arguments

video::Video: The video which defines the dimensions of the output

objects::Vector{Object}: All objects that are performed

frame::Int: Specific frame to be returned

Returns

Array{ARGB32, 2} - request frame as a matrix

Javis.interpolation_to_transition_val — Method

interpolation_to_transition_val(interpolation_val, Transition)

Returns the transition value for the given interpolation_val. If the interpolation value is already of the correct form it just gets returned. Otherwise the Transition function like get_position is called and the interpolated value is calculated.

Javis.match_num_point! — Method

match_num_point!(poly_1::Vector{Point}, poly_2::Vector{Point})

This is a helper function for morph. Given two polygons poly_1 and poly_2 points are added to the polygon with less points until both polygons have the same number of points. The polygon with less points gets mutated during this process.

Arguments

poly_1::Vector{Point}: The points which define the first polygon

poly_2::Vector{Point}: The points which define the second polygon

Javis.path_move — Method

path_move(x,y)

Moving to the specified point

Javis.path_quadratic — Method

path_quadratic(c_pt::Point, x,y, xe, ye)

Drawing a quadratic bezier curve by computing a cubic one as that is supported by Luxor

Javis.pathsvg — Method

pathsvg(svg)

Convert an svg to a path using Luxor. Normally called via the latex command. It handles only a subset of the full power of svg.

Javis.save_morph_polygons! — Method

save_morph_polygons!(object::AbstractObject, from_func::Function, to_func::Function)

Converts the paths created by the functions to polygons and calls match_num_point! such that both polygons have the same number of points. This is done once inside _morph. Saves the two polygons inside object.opts[:from_poly] and object.opts[:to_poly].

Assumption: Both functions create only a single polygon each.

Javis.set_attr — Method

set_attr(::Val{:transform}, transform_str)

Call the corresponding set_transform method i.e matrix, scale and translate

Javis.set_attrs — Method

set_attrs(o)

Setting the attributes of the object o by calling set_attr methods.

Javis.set_frames! — Method

set_frames!(parent, elem, last_frames::UnitRange; is_first=false)

Compute the frames based on a.frames and last_frames. Save the result in a.frames.frames which can be accessed via get_frames.

Arguments

parent is either nothing or the Object for the Action

elem is the Object or Action

last_frames holds the frames of the previous object or action.

is_first defines whether this is the first child of the parent (for actions)

Javis.set_object_defaults! — Method

set_object_defaults!(object)

Set the default object values

line_width and calls Luxor.setline.

opacity and calls Luxor.opacity.

scale and calls Luxor.scale.

Javis.set_transform — Method

set_transform(::Val{:matrix}, args...)

Multiply the new matrix with the current matrix and set it.

Javis.update_ObjectSetting! — Method

update_ObjectSetting!(as::ObjectSetting, by::ObjectSetting)

Set the fields of as to the same as by. Basically copying them over.